Snapshot progress.
This is a working version (for me at least) that sources all data from a
local zip file, and does not use any props file to store state.  Caveats
apply because the code was slightly hacked up to make this work, country
codes are still missing, and the code is crap.  But, it works.