Add Id to all blocks
The Id property is used to uniquely identify blocks in a PDF.  They
increment monotonically starting at zero, but there is no guarantee that
they are linear only that they are unique.  (Blocks may be merged which
breaks the linear guarantee.)

Replace a test case that just counts blocks with an Approval test.  This
is a much stronger guarantee, and completely locks down the serialized
form of a PDF document.  I should have done this a long time ago.