Start of BlockStateFactory
All features require knowledge of block, line, font, and font size
transistions.  This knowledge requires metadata about how the document is
parsed from top to bottom.  The rest of the features are based on token's
text, and require no document knowledge.  BlockStateFactory is used to
produce this document feature knowledge.

A few test cases have been sketched out, and are not passing.  The font
status tests are implemented, and passing.