Fixes for various feature vector screwups
The names of the some of the enums was incorrect.  This was immediately
obvious when diff'ing GROBID's output vs. GROBID.NET's approval.

Add lower case feature vector.
Fix suffix computation.

There are still issues the comparison of floating point values.  A
tolerance must be taken into account.

When TokenBlocks are tokenized this tokenization must be propigated up to
TextBlock and Block.