Implement TrainerRow
TrainerRow is a feature row returned by a CRP engine that contains the
expected value and the obtained value.  The expected value is extracted
from a known good source (TEI header).  The obtained value is calculated
based on the features of the row and a trained model.

A trainer row is similar to the following

 The the 0 1 0 1 0 BLOCKIN LINESTART 0 1 <title> <title>

A trainer row is almost identical to a feature row (and duplicates most of
the code).  A feature row has one classification whereas a trainer has
two.

Consider ways to merge these classes.